Algorithms run in own threads, algorithms optimized and remove paint
Algorithms now run in their own threads and have been optimized by painting just a single square instead of repainting the map. This alowed me to remove the pain method all together which made it so efficient I had to add a sleep to slow down the speed of the algorythm. looking into ScheduledExecutorService for next update